+++ src/views/clusteroverview/SettingsTab/UserTab/components/GettingStartedSection/GettingStartedSection.tsx
@@ -4,7 +4,7 @@ import {
   runningTourSignal,
   startTour,
   stopTour,
-} from '@kubevirt-utils/components/GuidedTour/utils/constants';
+} from '@kubevirt-utils/components/GuidedTour/utils/guidedTourSignals';
 import { useKubevirtTranslation } from '@kubevirt-utils/hooks/useKubevirtTranslation';
 import useKubevirtUserSettings from '@kubevirt-utils/hooks/useKubevirtUserSettings/useKubevirtUserSettings';
 import { USER_TAB_IDS } from '@overview/SettingsTab/search/constants';
