+++ src/views/welcome/components/WelcomeButtons.tsx
@@ -1,7 +1,7 @@
 import React, { FC } from 'react';
 import { useNavigate } from 'react-router-dom-v5-compat';
 
-import { startTour } from '@kubevirt-utils/components/GuidedTour/utils/constants';
+import { startTour } from '@kubevirt-utils/components/GuidedTour/utils/guidedTourSignals';
 import { ALL_NAMESPACES, ALL_NAMESPACES_SESSION_KEY } from '@kubevirt-utils/hooks/constants';
 import { useKubevirtTranslation } from '@kubevirt-utils/hooks/useKubevirtTranslation';
 import { useActiveNamespace } from '@openshift-console/dynamic-plugin-sdk';
